requirements.txt
asgiref==3.8.1
Django==5.0.7
django-filter==24.3
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
